+++ app/uk/gov/hmrc/tradergoodsprofilesrouter/models/request/UpdateRecordRequest.scala
@@ -36,7 +36,7 @@ case class UpdateRecordRequest(
   countryOfOrigin: Option[String] = None,
   category: Option[Int] = None,
   assessments: Option[Seq[Assessment]] = None,
-  supplementaryUnit: Option[Int] = None,
+  supplementaryUnit: Option[BigDecimal] = None,
   measurementUnit: Option[String] = None,
   comcodeEffectiveFromDate: Option[Instant] = None,
   comcodeEffectiveToDate: Option[Instant] = None
@@ -52,7 +52,7 @@ object UpdateRecordRequest {
       (JsPath \ "countryOfOrigin").readNullable(lengthBetween(1, 2).andKeep(verifying(isValidCountryCode))) and
       (JsPath \ "category").readNullable[Int](verifying[Int](category => category >= 1 && category <= 3)) and
       (JsPath \ "assessments").readNullable[Seq[Assessment]] and
-      (JsPath \ "supplementaryUnit").readNullable[Int] and
+      (JsPath \ "supplementaryUnit").readNullable[BigDecimal] and
       (JsPath \ "measurementUnit").readNullable(lengthBetween(1, 255)) and
       (JsPath \ "comcodeEffectiveFromDate").readNullable[Instant] and
       (JsPath \ "comcodeEffectiveToDate").readNullable[Instant])(UpdateRecordRequest.apply _)
